Move the Term Flow functions to AppLovinMAX from Privacy
