+++ src/types/AppLovinMAX.ts
@@ -1,5 +1,5 @@
 import type { Configuration } from './Configuration';
-import type { CmpError } from '../AppLovinMAX';
+import type { ConsentFlowUserGeography, CmpError } from '../AppLovinMAX';
 
 /**
  * Represents the AppLovinMAX module.
@@ -91,6 +91,45 @@ export type AppLovinMAXType = {
      * @param enabled Defaults to true.
      */
     setLocationCollectionEnabled(enabled: boolean): void;
+
+    /**
+     * @deprecated Use {@link setTermsAndPrivacyPolicyFlowEnabled()} instead.
+     *
+     * Enables the MAX Terms Flow.
+     *
+     * @param enabled true to enable the MAX Terms Flow.
+     */
+    setConsentFlowEnabled(enabled: boolean): void;
+
+    /**
+     * Enables the MAX Terms and Privacy Policy Flow.
+     *
+     * @param enabled true to enable the MAX Terms and Privacy Policy Flow.
+     */
+    setTermsAndPrivacyPolicyFlowEnabled(enabled: boolean): void;
+
+    /**
+     * The URL of your company’s privacy policy, as a string. This is required in order to enable
+     * the Terms Flow.
+     *
+     * @param urlString The URL string to point your company’s privacy policy.
+     */
+    setPrivacyPolicyUrl(urlString: string): void;
+
+    /**
+     * The URL of your company’s terms of service, as a string. This is optional; you can enable
+     * the Terms Flow with or without it.
+     *
+     * @param urlString The URL string to point your company’s terms of service.
+     */
+    setTermsOfServiceUrl(urlString: string): void;
+
+    /**
+     * Set debug user geography. You may use this to test CMP flow by setting this to {@link ConsentFlowUserGeography.GDPR}.
+     *
+     * @note The debug geography is used only when the app is in debug mode.
+     */
+    setConsentFlowDebugUserGeography(userGeography: ConsentFlowUserGeography): void;
 
     /**
      * Shows the CMP flow to an existing user.
